module breakdown (part three)
