+++ src/filters/namesFilterDict.js
@@ -1,0 +1,37 @@
+export const NamesFilterDict = [
+  {
+    id: `filter__all`,
+    label: `ALL`,
+    checked: true
+  },
+  {
+    id: `filter__overdue`,
+    label: `OVERDUE`,
+    checked: false
+  },
+  {
+    id: `filter__today`,
+    label: `TODAY`,
+    checked: false
+  },
+  {
+    id: `filter__favorites`,
+    label: `FAVORITES`,
+    checked: false
+  },
+  {
+    id: `filter__repeating`,
+    label: `REPEATING`,
+    checked: false
+  },
+  {
+    id: `filter__tags`,
+    label: `TAGS`,
+    checked: false
+  },
+  {
+    id: `filter__archive`,
+    label: `ARCHIVE`,
+    checked: false
+  }
+];
